requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0

# Database - Using psycopg3 for Python 3.14 compatibility
psycopg[binary]>=3.3.2

# Environment & Config
python-decouple==3.8

# CORS
django-cors-headers==4.3.1

# API Documentation
drf-spectacular==0.27.0

# Phone Number Validation
django-phonenumber-field[phonenumbers]==7.3.0

# OTP & Authentication
pyotp==2.9.0

# Redis & Celery (for OTP)
redis==5.0.1
celery==5.3.6

# Utilities
pillow>=10.3.0
python-dateutil==2.8.2

# Development
django-debug-toolbar==4.2.0

# Production Server
gunicorn==21.2.0